load beat map from s3
